feat: 자기소개 섹션 공유 프로필 추가
