+++ src/pages/profile/ProfilePage/components/IntroduceSection.tsx
@@ -1,52 +1,21 @@
-import { Flex, Heading } from '@/components';
-import { boxShadow } from '@/styles/common';
-import { styled, useTheme } from '@mui/material';
+import { styled } from '@mui/material';
 
+import { IntroduceContent } from '../../components/IntroduceContent';
+import { SectionBox } from '../../components/SectionBox';
 import { useGetProfileQuery } from '../../hooks/useGetProfileQuery';
 
 export const IntroduceSection = () => {
-  const theme = useTheme();
   const { profile } = useGetProfileQuery();
 
   return (
-    <S.Section
-      width="100%"
-      justify="center"
-      padding="1rem"
-      gap="1rem"
-      backgroundColor={theme.palette.variant.default}
-      wrap="wrap"
-      margin="auto 0"
-    >
-      <Heading as="h2" color={theme.palette.primary.main}>
-        About Me
-      </Heading>
-
-      <S.Content align="center" gap="1rem">
-        <S.Bar />
-        {profile?.self_description}
-      </S.Content>
-
+    <SectionBox height="150px">
+      <IntroduceContent introduce={profile.self_description} />
       <S.EditButton onClick={() => {}}>편집하기</S.EditButton>
-    </S.Section>
+    </SectionBox>
   );
 };
 
 const S = {
-  Section: styled(Flex.Column)`
-    border-radius: 1rem;
-    ${boxShadow}
-    position: relative;
-  `,
-  Content: styled(Flex.Row)`
-    height: fit-content;
-  `,
-  Bar: styled('div')`
-    background-color: ${({ theme }) => theme.palette.primary.light};
-    border-radius: 0.25rem;
-    height: 50px;
-    width: 10px;
-  `,
   EditButton: styled('button')`
     border: none;
     color: ${({ theme }) => theme.palette.primary.main};
